Move functionality into default plugin
